new: added DaxBaseImpl
